# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Assisted Migration Agent API
  version: v1
servers:
  - url: /api/v1
paths:
...
  /vms/{id}/utilization:
    get:
      summary: Get utilization breakdown for a specific VM
      operationId: getVMUtilization
      parameters:
        - name: id
          in: path
          required: true
          description: VirtualMachine MoRef ID
          schema:
            type: string
      responses:
        '200':
          description: VM utilization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmUtilizationDetails'
        '404':
          description: No utilization data found for this VM
        '500':
          description: Internal server error
...
  /rightsizing:
    post:
      summary: Trigger rightsizing metrics collection
      operationId: triggerRightsizingCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RightsizingCollectRequest'
      responses:
        '202':
          description: Collection triggered; returns report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingReportSummary'
        '400':
          description: Invalid request
        '500':
          description: Internal server error
    get:
      summary: List all rightsizing reports
      operationId: listRightsizingReports
      responses:
        '200':
          description: List of rightsizing reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingReportListResponse'
        '500':
          description: Internal server error
  /cluster_rightsizing:
    get:
      deprecated: true
      summary: Get cluster-level utilization aggregates for the latest completed report
      operationId: getLatestRightsizingClusters
      parameters:
        - name: byExpression
          in: query
          required: false
          description: Filter DSL expression (e.g. "cluster_id = 'domain-c123'"). Omit to return all clusters.
          schema:
            type: string
            maxLength: 10240
      responses:
        '200':
          description: Cluster utilization aggregates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingClusterListResponse'
        '404':
          description: No completed rightsizing report found
        '500':
          description: Internal server error
  /cluster_rightsizing/{report_id}:
    get:
      deprecated: true
      summary: Get cluster-level utilization aggregates for a specific report
      operationId: getRightsizingReportClusters
      parameters:
        - name: report_id
          in: path
          required: true
          description: Rightsizing report UUID
          schema:
            type: string
        - name: byExpression
          in: query
          required: false
          description: Filter DSL expression (e.g. "cluster_id = 'domain-c123'"). Omit to return all clusters.
          schema:
            type: string
            maxLength: 10240
      responses:
        '200':
          description: Cluster utilization aggregates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingClusterListResponse'
        '404':
          description: Report not found
        '500':
          description: Internal server error
  /rightsizing/{report_id}:
    get:
      summary: Get a specific rightsizing report with full VM metrics
      operationId: getRightsizingReport
      parameters:
        - name: report_id
          in: path
          required: true
          description: Report UUID
          schema:
            type: string
      responses:
        '200':
          description: Rightsizing report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingReport'
        '404':
          description: Report not found
        '500':
          description: Internal server error
  /rightsizing/{report_id}/clusters:
    get:
      summary: Get cluster-level utilization aggregates for a specific report
      operationId: listRightsizingReportClusters
      parameters:
        - name: report_id
          in: path
          required: true
          description: Rightsizing report UUID
          schema:
            type: string
        - name: byExpression
          in: query
          required: false
          description: Filter DSL expression (e.g. "cluster_id = 'domain-c123'"). Omit to return all clusters.
          schema:
            type: string
            maxLength: 10240
      responses:
        '200':
          description: Cluster utilization aggregates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingClusterListResponse'
        '404':
          description: Report not found
        '500':
          description: Internal server error
  /rightsizing/{report_id}/clusters/{cluster_id}:
    get:
      summary: Get utilization for a specific cluster from a specific report
      operationId: getRightsizingReportCluster
      parameters:
        - name: report_id
          in: path
          required: true
          description: Rightsizing report UUID
          schema:
            type: string
        - name: cluster_id
          in: path
          required: true
          description: vSphere MoRef (e.g. "domain-c123") or generated cluster ID (e.g. "cluster-a3f7b2c1d4e89012")
          schema:
            type: string
            pattern: '^(domain-c\d+|cluster-[0-9a-f]{16})$'
      responses:
        '200':
          description: Cluster utilization data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingClusterResponse'
        '404':
          description: Report not found, or cluster not found in report
        '500':
          description: Internal server error
  /clusters/{cluster_id}/utilization:
    get:
      summary: Get latest cluster utilization by cluster ID
      operationId: getClusterUtilization
      parameters:
        - name: cluster_id
          in: path
          required: true
          description: vSphere MoRef (e.g. "domain-c123") or generated cluster ID (e.g. "cluster-a3f7b2c1d4e89012")
          schema:
            type: string
            pattern: '^(domain-c\d+|cluster-[0-9a-f]{16})$'
      responses:
        '200':
          description: Cluster utilization data from the latest completed rightsizing report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingClusterResponse'
        '400':
          description: Invalid cluster_id format
        '404':
          description: No completed rightsizing report found, or cluster not found in report
        '500':
          description: Internal server error
...
components:
  schemas:
...
    VcenterCredentials:
      required:
        - url
        - username
        - password
      properties:
        url:
          type: string
          format: uri
          description: vCenter URL
          x-oapi-codegen-extra-tags:
            binding: "required,url"
        username:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        password:
          type: string
          format: password
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        cacert:
          type: string
          description: >-
            PEM-encoded CA certificate bundle for verifying the vCenter TLS
            certificate. Mutually exclusive with skipTls.
          x-oapi-codegen-extra-tags:
            binding: "omitempty"
        skipTls:
          type: boolean
          description: >-
            When true, TLS certificate verification is skipped. Must not be set
            to true alongside cacert — that combination returns 400. When
            omitted: if cacert is absent the connection skips TLS verification
            (backwards compatibility); if cacert is provided the connection
            verifies TLS using that certificate — skipTls need not be sent.
...
    RightsizingCollectRequest:
      type: object
      required:
        - credentials
      properties:
        credentials:
          $ref: '#/components/schemas/VcenterCredentials'
        lookback_hours:
          type: integer
          description: Lookback window in hours (default 720 = 30 days)
          default: 720
        interval_id:
          type: integer
          description: vSphere historical interval ID in seconds (300=day, 1800=week, 7200=month)
          default: 7200
        name_filter:
          type: string
          description: Filter VMs by name substring
        cluster_id:
          type: string
          description: MoRef value of a ClusterComputeResource to scope discovery (e.g. domain-c123)
        batch_size:
          type: integer
          description: Number of VMs per QueryPerf round-trip
          default: 64
        discover_vms:
          type: boolean
          description: >
            If true, discover VMs live from vSphere.
            If false (default), use the VMs already stored in the local inventory.
          default: false
    RightsizingMetricStats:
      type: object
      required:
        - sample_count
        - average
        - p95
        - p99
        - max
        - latest
      properties:
        sample_count:
          type: integer
        average:
          type: number
          format: double
        p95:
          type: number
          format: double
        p99:
          type: number
          format: double
        max:
          type: number
          format: double
        latest:
          type: number
          format: double
    RightsizingVMReport:
      type: object
      required:
        - name
        - moid
        - metrics
        - warnings
      properties:
        name:
          type: string
        moid:
          type: string
        metrics:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RightsizingMetricStats'
        warnings:
          type: array
          description: Non-empty when the VM was queried but had no historical metrics data.
          items:
            type: string
    RightsizingReport:
      type: object
      required:
        - id
        - vcenter
        - cluster_id
        - window_start
        - window_end
        - interval_id
        - expected_sample_count
        - vms
        - created_at
      properties:
        id:
          type: string
          description: UUID of the report
        vcenter:
          type: string
          description: vCenter URL used for this collection
        cluster_id:
          type: string
          description: MoRef of the ClusterComputeResource scoped for this collection (empty = all clusters)
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        interval_id:
          type: integer
        expected_sample_count:
          type: integer
          description: Theoretical maximum samples for the window (lookback / interval)
        vms:
          type: array
          items:
            $ref: '#/components/schemas/RightsizingVMReport'
        created_at:
          type: string
          format: date-time
    RightsizingReportSummary:
      type: object
      description: Rightsizing report metadata without VM metrics. Use GET /rightsizing/{id} to retrieve the full report including metrics.
      required:
        - id
        - vcenter
        - cluster_id
        - window_start
        - window_end
        - interval_id
        - expected_sample_count
        - created_at
      properties:
        id:
          type: string
          description: UUID of the report
        vcenter:
          type: string
          description: vCenter URL used for this collection
        cluster_id:
          type: string
          description: MoRef of the ClusterComputeResource scoped for this collection (empty = all clusters)
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        interval_id:
          type: integer
        expected_sample_count:
          type: integer
          description: Theoretical maximum samples for the window (lookback / interval)
        created_at:
          type: string
          format: date-time
    RightsizingReportListResponse:
      type: object
      required:
        - reports
        - total
      properties:
        reports:
          type: array
          description: List of report summaries. Use GET /rightsizing/{id} for full VM metrics.
          items:
            $ref: '#/components/schemas/RightsizingReportSummary'
        total:
          type: integer
          description: Total number of stored reports
    VmUtilizationDetails:
      type: object
      required:
        - moid
        - vm_name
        - provisioned_cpus
        - provisioned_memory_mb
        - provisioned_disk_kb
        - cpu_avg
        - cpu_p95
        - cpu_max
        - cpu_latest
        - mem_avg
        - mem_p95
        - mem_max
        - mem_latest
        - disk
        - confidence
      properties:
        moid:
          type: string
        vm_name:
          type: string
        provisioned_cpus:
          type: integer
          description: Number of vCPUs provisioned for this VM
        provisioned_memory_mb:
          type: integer
          description: Provisioned RAM (MB)
        provisioned_disk_kb:
          type: number
          format: double
          description: Provisioned disk (KB)
        cpu_avg:
          type: number
          format: double
          description: CPU utilization average (%)
        cpu_p95:
          type: number
          format: double
        cpu_max:
          type: number
          format: double
        cpu_latest:
          type: number
          format: double
        mem_avg:
          type: number
          format: double
          description: Memory utilization average (%)
        mem_p95:
          type: number
          format: double
        mem_max:
          type: number
          format: double
        mem_latest:
          type: number
          format: double
        disk:
          type: number
          format: double
          description: Disk utilization (%)
        confidence:
          type: number
          format: double
          description: Data confidence — sample_count / expected_sample_count × 100
    RightsizingClusterUtilization:
      type: object
      required:
        - cluster_id
        - cluster_name
        - vm_count
        - cpu_avg
        - cpu_p95
        - cpu_max
        - mem_avg
        - mem_p95
        - mem_max
        - disk
        - confidence
        - total_provisioned_cpus
        - total_provisioned_memory_mb
        - total_provisioned_disk_kb
      properties:
        cluster_id:
          type: string
          description: vSphere MoRef of the ClusterComputeResource (e.g. "domain-c123").
        cluster_name:
          type: string
          description: Display name of the cluster as reported by vCenter (e.g. "production-cluster").
        vm_count:
          type: integer
          description: Number of VMs with utilization data in this cluster
        cpu_avg:
          type: number
          format: double
          description: "SUM(vCPUs × cpu_avg%) / SUM(vCPUs) — weighted average CPU utilization (%)"
        cpu_p95:
          type: number
          format: double
          description: >
            SUM(vCPUs × vm_p95%) / SUM(vCPUs).
            Weighted summary of per-VM p95 values — not a true cluster-wide p95
            (which would require time-aligned samples).
        cpu_max:
          type: number
          format: double
          description: Weighted summary of per-VM max values (see cpu_p95 note)
        mem_avg:
          type: number
          format: double
          description: "SUM(provisionedMB × mem_avg%) / SUM(provisionedMB) — weighted average memory utilization (%)"
        mem_p95:
          type: number
          format: double
          description: Weighted summary of per-VM memory p95 values (see cpu_p95 note)
        mem_max:
          type: number
          format: double
          description: Weighted summary of per-VM memory max values (see cpu_p95 note)
        disk:
          type: number
          format: double
          description: "SUM(provisionedKB × disk%) / SUM(provisionedKB) — weighted average disk utilization (%)"
        confidence:
          type: number
          format: double
          description: >
            SUM(vCPUs × confidence%) / SUM(vCPUs).
            vCPU-weighted data coverage (0–100 %). Interpret utilization cautiously when low.
        total_provisioned_cpus:
          type: integer
          description: Sum of vCPU counts across all VMs in the cluster
        total_provisioned_memory_mb:
          type: integer
          description: Sum of provisioned RAM (MB) across all VMs in the cluster
        total_provisioned_disk_kb:
          type: number
          format: double
          description: Sum of provisioned disk (KB) across all VMs in the cluster
    RightsizingClusterListResponse:
      type: object
      required:
        - report_id
        - clusters
      properties:
        report_id:
          type: string
          description: The report ID this data was aggregated from
        clusters:
          type: array
          items:
            $ref: '#/components/schemas/RightsizingClusterUtilization'
    RightsizingClusterResponse:
      type: object
      required:
        - report_id
        - cluster
      properties:
        report_id:
          type: string
          description: The report ID this data was aggregated from
        cluster:
          $ref: '#/components/schemas/RightsizingClusterUtilization'
...